move the AC_CANONICAL_HOST and host specific flag tests into libreplace. This should fix the standalone build of tdb on HPUX, where we need to blacklist mmap.
Unfortunately this requires that we have a copy of config.guess and
config.sub in each of our project subdirectories. I tried to find a
way to use something like AC_CONFIG_AUX_DIR($libreplacedir) and just
put config.{guess,sub} in the lib/replace/ directory, but I couldn't
figure out how to do that in a way that kept autoconf happy for each
of our separate builds. Any autoconf guru out there see a way to do
this?


git-svn-id: svn://svnanon.samba.org/samba/branches/SAMBA_4_0/source/lib/talloc@18535 0c0555d6-39d7-0310-84fc-f1cc0bd64818